requirements.txt
fastapi-cloudauth==0.4.3
fastapi==0.115.4
uvicorn==0.32.0
httpx==0.27.2
curlify==2.2.1
python-dotenv==1.0.1
boto3==1.35.59
PyJWT[crypto]